# Drop anymarkup dependency, use LibYAML