--- a/dockerfiles/Dockerfile
+++ b/dockerfiles/Dockerfile
@@ -5,9 +5,9 @@ ENV LANG=en_US.utf8
 
 RUN yum install -y epel-release && \
     yum install -y python36 python36-pip && \
+    yum install -y python36-PyYAML && \
     yum clean all && \
-    pip3 install --upgrade pip && \
-    yum clean all
+    pip3 install --upgrade pip
 
 COPY validator /validator/validator
 COPY setup.py /validator