--- a/validator/test/fixtures.py
+++ b/validator/test/fixtures.py
@@ -1,6 +1,6 @@
 import os
 
-import anymarkup
+from validator.utils import parse_anymarkup
 
 
 class Fixtures(object):
@@ -20,4 +20,4 @@ class Fixtures(object):
             return f.read().strip()
 
     def get_anymarkup(self, fixture):
-        return anymarkup.parse(self.get(fixture), force_types=None)
+        return parse_anymarkup(self.path(fixture))