--- a/validator/test/fixtures.py
+++ b/validator/test/fixtures.py
@@ -1,6 +1,6 @@
 import os
 
-from validator.utils import parse_anymarkup
+from validator.utils import parse_anymarkup_file
 
 
 class Fixtures(object):
@@ -20,4 +20,4 @@ class Fixtures(object):
             return f.read().strip()
 
     def get_anymarkup(self, fixture):
-        return parse_anymarkup(self.path(fixture))
+        return parse_anymarkup_file(self.path(fixture))