--- a/validator/validator.py
+++ b/validator/validator.py
@@ -6,11 +6,12 @@ import sys
 from enum import Enum
 from functools import lru_cache
 
-import anymarkup
 import click
 import jsonschema
 from jsonschema import Draft6Validator as jsonschema_validator
 import requests
+
+from validator.utils import parse_anymarkup
 
 logging.basicConfig(format='%(levelname)s: %(message)s', level=logging.WARNING)
 
@@ -341,7 +342,7 @@ def fetch_schema(schema_url):
         r = requests.get(schema_url)
         r.raise_for_status()
         schema = r.text
-        return anymarkup.parse(schema, force_types=None)
+        return parse_anymarkup(schema)
     else:
         raise MissingSchemaFile(schema_url)