# Context
For a long time I've been seeing qontract-validator take longer and longer to run. After digging a bit, this appears to be priparily due to the fact we're using a version of PyYAML that is not compiled with LibYAML support. Furthermore, it is not possible to tell anymarkdown to use the C-based LibYAML loaders and dumpers even if they are available.

# PyYAML
We list PyYAML as a dependency. This makes sense as it is required to run the app. However in order to ensure it is built with LibYAML support, the client also needs to have the libyaml and python development libraries in addition to having a compiler and related build tools.

In order to achieve this for our needs, I've updated the Dockerfiles to install PyYAML from the repos. Python finds the system installed package before the user-installed one (it searches /usr/lib64 before /usr/local/lib64)

# Replacing anymarkup
The main change in this PR is dropping the anymarkup dependency in favor of our own markup parsing functions (see `parse_anymarkup`)

The initial implementation only accepts json and yaml because that is what we use right now. It should be easy to extend the functionnality to more markup languages if needed in the future.

The new functionallity will use the LibYAML `C` functions by default (`CSafeLoader`) and fall back to the native (and slower) implementation (`SafeLoader`) if LibYAML is not available.

# Tests

The test suite is still reporting success after this change.

I have also compared generated bundles from before & after these changes and the resulting bundles are identical 

Some rudimentary performance testing has shown the following results:
| Test   | Before   | After  | Improvement |
| -------|----------|--------|-------------|
| local  | 10s      | 1.5s    | 85%         |
| docker | 20s       | 5s      | 75%         |
| podman | 20s       | 4s      | 85%         |