@rporres 2021-01-13
Great stuff, @jfchevrette. The improvement is massive and I like the `utils.py` approach. I've compared the md5 sums of a bundle created with the current code and yours and they match. Just a small thing to be corrected in the json loading to make the validator work. Once there, feel free to merge this.